LayoutTests/fast/canvas/CanvasRenderingContext2DSettings-pixelFormat-enabled-expected.txt

Test that the pixelFormat member of CanvasRenderingContext2DSettings is supported when CanvasPixelFormatEnabled=true.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testSettings.pixelFormat is "uint8"
PASS testSettings.pixelFormat is "uint8"
PASS testSettings.pixelFormat is "float16"
PASS document.createElement("canvas").getContext("2d", { pixelFormat: "foo" }) threw exception TypeError: Type error.
PASS successfullyParsed is true

TEST COMPLETE
